indexer-alt: detect duplicate pipelines added
## Description

Catch issues like the one we had with `kv_feature_flags` which was
incorrectly named `kv_protocol_configs` (copypasta error). This will now
fail when the pipeline is added the second time (by name).

## Test plan

Recreate the scenario with `kv_feature_flags` -- the indexer will panic
before it starts in that case.