indexer-alt: generalize GraphQLConfig
## Description

`GraphQLConfig` is a derive macro for deriving `Serialize`,
`Deserialize`, and `Debug`, and introducing `#[serde(default = ...)]`
annotations that rely on the fields of the type's `Default`
implementation.

This allows config types to rely on their `Default` implementation as a
single source of truth for default values, and is used through-out
`sui-graphql-rpc` to define its config types.

This change generalizes it (renames it) so it can be used in the indexer
as well.

## Test plan

Build and test `sui-graphql-rpc`, `sui-mvr-graphql-rpc`:

```
sui$ cargo nextest run -p sui-graphql-rpc
sui$ cargo nextest run -p sui-graphql-e2e-tests
sui$ cargo nextest run -p sui-mvr-graphql-rpc
```

This is a behaviour preserving change for the moment.