indexer-alt: DbConfig::connection_timeout_ms
## Description

Change `DbConfig::connection_timeout: Duration` to
`DbConfig::connection_timeout: u64`, so we can drop the dependency on
`const_str`.

Also introduce a default value for the database URL so that we can
implement a `Default` impl for `DbConfig` overall, and which serves as a
single source of truth for defaults.

## Test plan

```
sui$ cargo run -p sui-indexer-alt -- --help
```