fix(sui): be consistent about sui config directory
## Description
When a `--config` is supplied and it is a directory, we treat it as the
sui config directory in some but not all cases. This changes makes it so
that if `--config` does not look like a YAML file, then it is treated
like it should be the config directory, and that location is used
everywhere the config directory is expected.

## Test plan

Manually tested running `sui start`, `sui start --config ...`.