indexer-alt: support for bulk update
## Description

Although postgres supports bulk-updating rows using `VALUES`, diesel
does not natively support it. This change adds support for this. It is
in its own crate so that we can limit the fallout of depending on the
diesel breaking changes feature (which we need to depend on to access
the types that diesel converts collections of model types into, ready to
be inserted into the DB).

## Test plan

Unit tests for the query generation, and E2E tests for running updates
on a DB with this new DSL:

```
sui$ cargo nextest run -p diesel-update-from
```