refactor(indexer-alt): remove dependency on mysten-metrics
## Description
Part of the preparation for factoring out the indexer framework and
schema. Removing a dependency on `mysten-metrics` because this is
probably not something we want people outside of Mysten to depend on,
and it is used in code that would be factored into the indexer
framework.

This means we lose some metrics related to thread stalls, and number of
tracked tasks, but as far as I'm aware, we don't rely on these metrics
anywhere.

## Test plan
CI