refactor(indexer-alt): remove dependency on mysten-metrics (#20499)
## Description
Part of the preparation for factoring out the indexer framework and
schema. Removing a dependency on `mysten-metrics` because this is
probably not something we want people outside of Mysten to depend on,
and it is used in code that would be factored into the indexer
framework.

This means we lose some metrics related to thread stalls, and number of
tracked tasks, but as far as I'm aware, we don't rely on these metrics
anywhere.

## Test plan
CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: