[move-ide] Track all files requiring analysis (#20510)
## Description 

Currently move-analyzer does not fully track all files requiring
analysis but rather only the last one sent to the symbolicator. It
should not be a huge deal from the user's point of view as they spend
significant amount of time (at least compared to analysis time) editing
a single file before switching to the next, but precise tracking will be
needed if we only want to re-analyze specific files (as part of an
upcoming optimization). This PR implements tracking of all files
requiring analysis.

## Test plan 

All tests must pass