rpc: convert Client error to be a tonic::Status
Convert Client error type to be a tonic::Status. In addition, ensure
that if http headers (which are wrapped in a
tonic::metadata::MetadataMap) are received, that they are included in
either the response or the error Status so that callers can retrieve
header information even in the event of an error.