fix(sui-indexer): allow setting --gc-checkpoint-files=false
Previously I was unable to turn off checkpoint file garbage collection,
since the default value was `true`, the `--gc-checkpoint-files` flag
didn't expect any values and the default `clap` action was to set the
value to `true` when the flag was present. I've tried all of:

- `--gc-checkpoint-files 0`
- `--gc-checkpoint-files false`
- `--gc-checkpoint-files=0`
- `--gc-checkpoint-files=false`

This changes that so that the `false` value is accepted by the argument.
It is based on this [comment](https://github.com/clap-rs/clap/issues/1649#issuecomment-2144879038)