[docs][2/2] Update deepbook code references (#20770)
## Description 

Updates code to be reference. Relies on
https://github.com/MystenLabs/sui/pull/20701.

This did not work as I had planned. The only review needed here is the
code reference changes to the deepbook content. The rest is part of a
different PR.

## Test plan 

Local

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: