telemetry-subscribers: remove feature flags (#20820)
Remove the feature flags from the `telemetry-subscribers` crate. The
existing flags were not being properly handled which could have lead to
broken compilation if features were enabled/disabled and given we don't
make use of feature flags in the sui repo it makes things easier to
maintain with these removed.