[indexer-alt] Remove coin balance bucket pruner (#20805)
## Description 

This PR reimplements the coin balance bucket pruner using proper pruner
instead of a pipeline.
This is the last pipeline that uses consistency layer. So we also get
rid of all the related implementation code for it.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: