sui-rpc: rework proto files based on 'buf lint'
Rework the sui package proto files based on suggestions from 'buf lint'.
One of the suggestions is to ensure that services end with a common
prefix `Service`. In order to gracefully transition to this for the
release to mainnet, the renamed service `NodeService` and the old
service `Node` will both be supported for a single release, after which
the older `Node` service will be removed.