[Consensus] Linearizer to use commit up to gc_round (#20250)
## Description 

Currently when committing/collecting the sub dag we always stop the
recursion when we come across the highest committed round for an
authority. This works well, but with the upcoming Fast Path changes it
can lead to uncommitted FP certificates in the presence of blocks that
don't form a proper hash chain.

This could be done if for example an authority `A` produces blocks
`A100, A101, A102` , but block `A102` uses as ancestor block `A100` and
never references `A101`. `A101` gets broadcasted and participate in the
form of an FP certificate and referenced by others. With the current
logic in Linearizer, if `A102` gets committed first, then `A101` will
never get committed as part of subsequent DAG sequencing as we always
stop our recursion to the highest committed round. With the new logic we
ensure that `A101` will get committed as well if it's `> gc_round`.
Having `Garbage Collection` allows us to limit the recursion depth
making the process efficient enough and knowing that we'll not attempt
to keep committing blocks too far in the past. Using a reasonable depth
(ex 50 - 60) for production shouldn't cause performance troubles.

To achieve the above `DagState` has been refactored to keep a global
commit state of blocks. When we linearise the DAG across commits we do
not re-commit previously committed blocks.

The new changes are guarded behind a feature flag and can only be
enabled if GC is enabled.

## Test plan 

CI/PT

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [x] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: