indexer-alt: configure consistency retention in one place
## Description

Add back a dedicated top-level configuration for consistency, to control
the retention of the consistent range. Now that this range is controlled
by the pruner, it can be configured by a `PrunerConfig` as well, rather
than a dedicated `ConsistencyConfig`.

The motivation for this is that the E2E testing setup needs to configure
the consistent range, and it would be better to avoid having to
enumerate all the consistent pipelines in the test harness because it
would be easy to forget to update that list, whereas with this approach,
we take advantage of the fact that there is an `add_consistent` macro to
register the pipeline, and it will do the right thing.

## Test plan

```
sui$ cargo check
sui$ cargo nextest run -p sui-indexer-alt
sui$ cargo run -p sui-indexer-alt -- generate-config
```