indexer-alt: add tx_digest -> tx_sequence_number index
## Description

Add an index on `tx_digests` from digest to sequence number (the reverse
of the primary key index). This is to make it easier to look up balance
changes (keyed by sequence number in the database) based on their digest
(the key we use for the KV store).

The alternative was to make `tx_balances` keyed on digest, but in that
case we would still need to retain the sequence number, because it is
used for pruning.

## Test plan
CI (Ran `./generate_schema.sh`)