fix(open-rpc): Handle empty doc comment lines
## Description

The code that extracts doc comments to generate an OpenRPC schema panics
on empty lines. This change fixes that, treating an empty line as a hard
line break.

## Test plan

Start the new JSONRPC implementation, and inspect its output which
contains the OpenRPC schema it offers:

```
sui$ cargo run -p sui-indexer-alt-jsonrpc
```

A follow-up change introduces a doc comment that includes empty lines.