jsonrpc: unlimit the number of concurrent jsonrpc requests
The jsonrpsee library is artificially limiting the number of concurrent
requests that the server can process to ~100 since its treating each
request as a separate connection. In order to work around this we can
just set the maximum number of connections for the jsonrpsee service to
a high number since its not actually controlling the number of
connections.