Add `ExecutionTimeEstimate` mode for congestion control (#20994)
## Description 

- `ExecutionTimeObserver` tracks local measurements of execution time
and submits observations to consensus when moving average changes by
more than a threshold.

- `ExecutionTimeEstimator` records execution time observations received
from consensus and computes stake-weighted medians for use in congestion
control.

This PR contains the core implementation of the feature but is missing
some important components required for use in production:
- storage of received observations for crash recovery
- propagation of observations to next epoch
- metrics

The heuristics it uses are as simple as possible for a first pass and
will likely require some tuning.

## Test plan 

Added & updated unit tests. Disabled for now by protocol config.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: