[1/n][vm-rewrite][Move] Updates to VM: some bugfixes, and query functions on vm instances needed for the adapter.
This adds various "query" functions that the Sui adapter will need to
the VM (e.g., resolving a typetag to a loaded type, getting type
layouts, abilities, etc). As part of this it adds a
`LoadedFunctionInformation` struct that allows VM runtime `Type`s to
flow out of the VM.  The adapter and object runtime will need to (and
does) take care to not allow these VM runtime types to flow across VM
instance boundaries.

The code in the PR may not be working as future PRs will build on top of
this.