[vm-rewrite][sui][roughcut] Update sui deps on Move and add new trait implementations.
Things are still a bit rough around the edges in terms of getting the
execution versioning story right around deps into the VM-runtime (e.g.,
for `Serializable` value, cost tables etc). But that will be done in a
later PR. There is also the question of how we want to handle the
different `GaMeter` traits, but again -- to he handled later in a
separate PR. We probably want to discuss this a bit before deciding what
to do here as well as there are a couple different options available to
us.