[5/n][vm-rewrite][sui][roughcut] Update sui deps on Move and add new trait implementations.
Things are still a bit rough around the edges in terms of getting the
execution versioning story right around deps into the VM-runtime (e.g.,
for `Serializable` value, cost tables etc). But that will be done in a
later PR.

In particular, here the direct dependency on
`move-execution/crates/shared` in sui-types will be removed once we pull
the shared legacy code crate out of `move-execution` and into something
like `legacy_vm_types` or similar under `move/crates`.

NB: The code in the PR may not be working as future PRs will build on top of
this.