[7/n][vm-rewrite][sui-execution] Update adapter
This is the big one...

There is still some refactoring and changes need, but broad-strokes this
plumbs in the new VM to the adapter.

The main conceptual change is the idea of a `LinkedContext` which is an
execution context with a specific linkage and VM instance associated
with it. All commands currently require a `LinkedContext`.

On the TODO list is:
1. To relax the need for a `LinkedContext` for `Upgrade` and `Publish`
   commands to require a context a priori (however there are some
   mechanics that will need to be sorted out to make this work).
2. Properly handle type versioning/defining IDs for typetags/type
   resolution.
3. General cleanup.

NB: The code in the PR may not be working as future PRs will build on top of
this.