Make transaction execution sync
This requires changing MutexTable to be sync, which in turn ripples out
into making a few other pieces of code sync, but the blast radius is
pretty well contained.

I do not expect this to have any significant performance impacts,
because most of the previous await points were probably "fake" in
that they were unlikely to actually yield execution.