[CLI] Don't error if we cannot connect to active env (#21116)
## Description 

If the CLI cannot get the client that should connect to the active
environment set in the config, it will error. This creates troubles for
when you work on a local network and then stop it. Every client (and a
few other) command will now error.

This PR fixes this by relaxing a bit the condition.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [x] CLI: Fixed the bug when the CLI cannot connect to the active
environment leading to an error on most commands.
- [ ] Rust SDK: