[CLI] store framework versions in sui binary (dvx-195) (#21018)
## Description 

This PR adds a module `sui::framework_version` containing functions to
obtain metadata about historical framework versions (git commit hashes,
system package names and paths). This information will be needed to
support implicit framework dependencies.

This information is gathered at build time from the manifest in the
`sui-framework-snapshot` crate. This required adding additional data to
that manifest, since the package names and source paths were not
previously stored. The code in `sui-framework-snapshot` that maintains
that manifest was updated.

## Test plan 

There are unit tests for `sui::framework_versions` that check the
correctness of the metadata that is returned. There are also updated
integration tests in `sui_framework_snapshot` that check that the
updated manifest matches the `Move.toml` files in
`sui-framework/packages`.