[12/n][vm-rewrite] Update and cleanup linkage generation
This updates linkage generation to almost always include system packages
(other than in genesis/system transaction mode -- so basically when
we're updating the system packages). This is primarily since we rely on
access to these packages at places in the adapter (e.g., when resolving
the type for the upgrade capability when upgrading or publishing a
package).

This also cleans up the interfaces around linkage