[framework-snaps] Update snapshot loading (#21193)
Since deployed protocol versions are not guaranteed to be sequential,
snapshots are not guaranteed to be sequential (e.g., 45 may exist, 46
may not, but 47 may).

Previously when trying to load a snapshot for a protocol version we
would only load a snapshot if there was one for that exact version,
otherwise we'd default to the newest version of the framework. This is
incorrect, and @dariorussi has run into problems with this.

This updates snapshot loading so if a snapshot for a particular version
`v` is not found, the least version greater than the given version is
returned.

## Test plan 

CI
 
---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------

Co-authored-by: ronny-mysten <118224482+ronny-mysten@users.noreply.github.com>