[Consensus] introduce consensus only simtests (#20683)
## Description 

PR is introducing the framework to enable us run Consensus-only
simtests. Looking on the existing sui-swarm it would be a bit
challenging to reuse as APIs are different and especially when we need
to access consensus specific components.

The PR is using similar patterns that have been used in sui-swarm but I
am expecting to refine this as we go. This might be a good start for
now. A simple simtest method has been added for the time being
`test_committee_start_simple` . Eventually we want to replace the tests
on `authority_node` and add a few more cases. Motivation to do this has
been the Garbage Collection work when investigating some edge cases
where reproducibility was important to confirm the issues.

Example to run this simtest with additional logging, a high regional
variance in latency, and outputting the result in `output.txt`

```
SUI_SIM_CONFIG=regional_high_variance RUST_LOG=info,consensus_core=debug cargo simtest test_committee_start  --nocapture > output.txt 2>&1
```

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: