rpc-alt: Macro for building diesel `SqlLiteral`s
## Description

Introduce a macro that allows for building a `SqlLiteral` using a
format-string-like syntax.

`SqlLiteral` should allow us to keep using `diesel` and not have to
worry about how to handle scenarios where we need to introduce snippets
of raw SQL -- it makes it possible to do so while still taking advantage
of strong typing on the boundary and diesel's support for escaping
values.

However, it has poor ergonomics, requiring lots of boilerplate (chains
of function calls and type annotations) which break up and hide the
query being built. The newly introduced macro allows for building the
same structure using a format string. See diff for docs and examples of
usage.

Some features that were not included in the initial implementation, but
could be include:

- inlining binds (similar to inlining interpolated values in format
  strings:
  - `sql!("foo = {bar:BigInt}")` instead of
  - `sql!("foo = {BigInt}", bar)`
- named binds (again, similar to the same feature in `format!` strings):
 -  `sql!("foo = {bar:BigInt}", bar = 1 + 2)`.
- support for normal format strings, which in particular would allow us
  to interpolate column and table names into sql snippets.

## Test plan

The macro comes with extensive unit tests, and the usages are covered by
existing tests:

```
cargo nextest run            \
  -p sui-sql-macro           \
  -p sui-indexer-alt         \
  -p sui-indexer-alt-jsonrpc \
  -p sui-indexer-alt-e2e-tests
```