rpc-alt: fix transactions next cursor behavior
## Description

If the transaction query returns an empty result and a cursor was
supplied, then return that cursor back in the response, so that the
client can poll for new results by repeatedly requesting new pages.

This matches the revised behaviour for pagination in `sui-indexer`'s
`reader`. Note that in `rpc-alt`, this behaviour is currently only
applicable to paginating transactions -- object and coin pagination flow
backwards through time, so if you reach an empty result, there is no
point retrying it.

## Test plan

```
sui$ cargo nextest run -p sui-indexer-alt-e2e-tests
```