:memo: docs: improve randomness-onchain (#21242)
## Description 

Describe the changes or additions included in this PR.

We update the documentation on how to use on-chain randomness. The
current documentation places a lot of emphasis on gas consumption, when
a similar attack can be done by hitting any resource limit.

## Test plan 

N/A

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------

Co-authored-by: ronny-mysten <118224482+ronny-mysten@users.noreply.github.com>