chore: Add cargo deny check for external crates (#21268)
Currently the external crates cargo deny job runs at the top level and
thus does not check the external crates manifest/lock file. This PR adds
an additional `deny.toml` specific to the external crates and updates
the CI job so that it uses the correct manifest. Also adds the
`--hide-inclusion-graph` flag as this helps to reduce the spam in the CI
logs.

NOTE: If only the top level `deny.toml` is used for both, then warnings
are emitted for ignored RUSTSEC, licenses, etc. that do not apply to the
manifest which cannot be silenced.