[vm-rewrite] Add ephemeral packages to linked context
In the case where a package init fails, we need access to the package
that failed to be published in order to resolve the error correctly.

This adds the concept of "ephemeral package" to the linked
context -- this is only for cases where you are doing something during
publication. If the publication is successful the package will appear in
the `new_packages` on the `Context`.