[json-rpc] Update variants to be returned in declaration order
We were collecting these into a `BTreeMap` when building up the json
response which was leading to the variants in an enum type to be
returned in lexicographic order instead of declaration order.

This updates this to collect into a vector of tuples instead so that
declaration order of the variants is preserved in the json output.