rpc-alt: Use `obj_versions` to determine liveness, instead of `obj_info`
## Description

Now that `obj_versions` includes rows when an object is deleted or
wrapped, it can be used directly to check for liveness. This fact can be
used to simplify live object fetching:

 - We no longer need to first query `obj_info` to see if the latest info
   suggests the object is live, and then get its latest version from
   `obj_versions` -- we can get the latest version directly and check
   whether it corresponds to a deleted/wrapped object.
 - This means the distinction between a "latest" and a "live" object no
   longer needs to exist (which gets rid of the potential for a bug
   where you fetch the latest version of an object when you meant to
   fetch the live version).
 - The `LatestObjectInfoKey` data loader can also now be deleted,
   because it doesn't offer anything over `LatestObjectVersionKey`.

This slightly undoes an optimisation that was introduced for
`LatestObjectVersionKey`, where it was optimised to only request columns
that were in the index being scanned (so the fetch could be an index
only scan). I have confirmed that after switching to `CROSS LATERAL
JOIN` this doesn't hurt performance (I suspect this is because when we
were using `SELECT DISTINCT ON`, it was loading data before filtering,
to sort, and it can avoid doing that entirely when we use `CROSS JOIN
LATERAL` so the number of heap pages it needs to fetch is now much
smaller).

The benefit is that we will no longer issue a fetch for an object from
the KV store that we know doesn't exist (because it has no digest in
`obj_versions`).

## Test plan

Existing tests:

```
sui$ cargo nextest run -p sui-indexer-alt-e2e-tests
```