Fix bug in cache coherency check found by cache size randomization
The previous code was not incorrect, because a tombstone must always be
treated identically to no-entry-found. But it presents a tricky edge
case for the coherency check that runs in debug mode.

The fix changes the cache to track tombstones if they exist, which is
also correct and will probably be less confusing in the future, and
which also makes the debug check work correctly.