[move] Added source file to debug info and used it to finesse trace debugger experience (#21495)
## Description 

This PR adds source file information to debug info for the file that
this info was generated for/from. Previously, dependency source files
used by the trace debugger are those in the `build` directory, as they
are easy to find and readily available. This however had two unfortunate
consequences for the trace debugger user:
- if during debugging, the user enters a function in a dependent source
file, they will not have access to Move IDE's code inspection
capabilities for this file (the reason for it is that source files in
the `build` directory are not "buildable" by the compiler due to lack of
the manifest file so the symbols required for on-hover, go-to-def etc.
cannot be produced)
- if the user uses go-to-def to go to a dependent source file to set a
breakpoint there, the breakpoint would not trigger (the reason for it is
that the dependent file reached this way is not the one in the `build`
directory that is known to the trace debugger but rather the actual
original dependent source file)

Having the original dependent file path in debug info allows us to
remedy these two shortcomings. Of course it is possible that this
original dependent file path no longer exists, in which case trace
debugger will default to using a file from the `build` directory.

## Test plan 

A new test for debug info v2 was added. All tests must pass. I also
tested manually that both deficiencies were remedies when using trace
debugger.