proto: rework proto messages based on feedback
This patch introduces a new protobuf package 'sui.rpc.v2beta' and
'sui.rpc.v2alpha' with completely reworked protobuf messages as well as
reworked rpc service definitions. In particular the changes to the
messages are as follows:

- Address and ObjectId are now represented as 'string' types using
  their canonical human-readable representation in hex.
- Digest is now represented as 'string' type using its
  canonical human-readable representation in base58.
- TypeTag and StructTag are now represented as 'string' types using
  their canonical human-readable representation.
- i128, u128, u256 and Bn254FieldElement are now represented as 'string'
  types using their canonical human-readable representation in base10.
- The Bcs type now includes a 'name' field that can be used to specify
  the type name of the serialized value.
- Various types representing a sui core type have been reworked and
  optimized for readability and accessibility over providing as close as
  possible a 1-1 mapping to their BCS representation.
- Many message definitions have been split into separate files for
  improved readability and maintainability.

This patch also implements and serves the new api and removes the old
sui.node.v2alpha.NodeService and sui.node.v2alpha.SubscriptionService
services. The sui.node.v2.NodeService service is left as-is and will be
removed in a subsequent release.