pick(21510): indexer-alt: obj_versions wrapped/delete sentinels
## Description

Record the version an object would have been at if it had not been
deleted/wrapped in `obj_versions`. This is required for implementing
child object fetches in GraphQL, but it also simplifies the logic of
reading live objects that might have been deleted for JSONRPC. Today
this process requires reading from `obj_info` to determine liveness, and
then looking up the latest version of the object in `obj_versions`. This
can be simplified to one query (to `obj_versions`) to get the latest
version, and also check whether it is live or not.

A special backfill pipeline has been added so that we can backfill rows
corresponding to wrapped/deleted sentinels without wiping the whole
table.

## Test plan

Existing tests.

## Stack

- #21471
- #21482
- #21487

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] gRPC:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [ ] Rust SDK: