Fix analytics pipeline initialization bug (#21579)
## Description

There is a long-standing off-by-1 bug when initializing a new pipeline.
The pipeline would be initialized to process checkpoint 1 first instead
of checkpoint 0 first in the case where there was no existing state for
the pipeline in the object store. This causes a panic which causes the
whole process to crash on release builds.

This PR just fixes the off-by-1 bug.

## Test plan

Manually verified the off-by-1 bug no longer occurs