Give each pipeline it's own checkpoint directory. (#21577)
## Description

This is fixing a bug in
[pull/21486](https://github.com/MystenLabs/sui/pull/21486). Each task
was sharing a "checkpoint_dir" which is where the output parquet/csv is
temporarily stored until it is sync'd in the background to the object
store. This could create problems if two pipelines for the same "type"
(objects, transactions, etc) are running in the same process since this
directory would end up conflicting (each pipeline creates a subdirectory
named after its type in the checkpoint_dir).

The fix is to create a unique temporary directory per task to ensure
there are no conflicts within a process.

Note that this PR motivated me to do [this subsequent
refactor](https://github.com/MystenLabs/sui/pull/21578) to clean up this
code a bit.

## Test plan

Manual testing.