fix(rpc-alt): fix CoinMetadata and Display data-loader ordering
## Description

Both these data-loaders follow a pattern where they generate some probes
into the DB, based on the data loader key, form a set of these to query
the DB with and then later try to zip that set with the original
sequence of keys they were generated from.

If that sequence of keys is in a different order or contains duplicates,
then keys are going to end up mis-aligned with their equivalent probes.

## Test plan

:eyes: