[CLI] Fix the --dump-bytecode-as-base64 to work with implicit deps (#21583)
## Description 

The `--dump-bytecode-as-base64` lives in the `sui_commands.rs` in the
CLI because of the need to pass in a client to have access to the
network. We forgot to add the implicit deps flag there, which this PR
fixes.

Should fix and close #21581

## Test plan 

Locally

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: