[CLI] removed scary external resolver output (#21742)
## Description 

Currently if an external resolver outputs anything it is marked as an
error; this PR removes that line. The output from the resolver is still
printed, and if the resolver exits with an error code, that is still
reported as an error.

## Test plan 

No additional tests; change is trivial.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [X] CLI: remove spurious error message for external resolvers
- [ ] Rust SDK: