indexer-alt: avoid panic in kv_protocol_configs and kv_feature_flags
## Description

Use `ProtocolConfig::get_for_version_if_supported` instead of
`ProtocolConfig::get_for_version`, so that if the version is not
supported, we don't panic and have a chance to propagate an error.

If this pipeline is running in the same process as other pipelines, this
will allow those other pipelines to continue making progress.

## Test plan

New unit tests for this behaviour.

```
sui$ cargo nextest run -p sui-indexer-alt
```