Fix flaky get_latest_object_version (#21828)
## Description 

This fixes a flaky simtest:
https://github.com/MystenLabs/sui/actions/runs/14418124154/job/40437321145
In some simtests we crash more than a quorum of validators, and in that
case we may not be able to get the object from a quorum of validators.
Adding retry here. This function is for testing only so we don't need to
be very strict.

## Test plan 

Reran simtest

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: