replay tool [1/N] (#21831)
## Description 

Initial work for the replay tool. 
This is a work in progress that used to be on a branch and we are moving
on main so other people (namely @awelc, @stefan-mysten, @tzakian, @amnn)
can easily integrate their work.
Not everything works yet, as the data is not yet always available, deny
list is not hooked up yet, execution tracing is not finished yet and
there are likely random bug.
Overall I think a cgood starting point as we continue to work on it

## Test plan 
No test plan, when we are going to plug in the cache we are going to
develop some tests

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: