[move] Added parsing resilience for control expressions to enable auto-completion (#21790)
## Description 

What the subject says. With this, we will have proper auto-completion
for incomplete `if`, `while` and `match` statements, for example:


![image](https://github.com/user-attachments/assets/790219fc-8f78-4abb-85ff-a51b1cf3e8c2)


## Test plan 

Output of compiler tests was adjusted (all tests must pass)

Auto-completion tests will be added once an agreement on the compiler
changes is reached

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [x] CLI: Compiler might generate slightly different errors as a result
of the compilation process being more permissive with respect to parsing
errors.
- [ ] Rust SDK: