Move sui metrics push client to a separate crate (#21874)
## Description 

mysten-common crate should not depend on sui-types. This is due to the
use of metrics push client.
This PR moves the metrics push client to a separate crate.
Also removes the duplicate code between sui-node and sui-bridge.
It also removes certificate from the client since it is not used.
Let me know if we still need it.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: