[rocks] Add option to debug log kv hash (#21863)
## Description 

This PR introduces a new option in DBOptions, when set, will debug log
the hash of both key and value on every insert and read.
This is then enabled for the full_checkpoint_contents table.
Also added the optimize_for_large_value option for this table since it
seems to make sense there and should help with performance.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: