graphql-alt: Object atCheckpoint constraint
## Description

Make it possible to fetch the latest version of an object as of a
particular checkpoint.

As of this change, this query pattern is only exposed through
`Query.object` and `Query.multiGetObjects` via an explicit
`atCheckpoint` parameter, but it will be used in a couple of places:

- `Query.object` and `Query.multiGetObjects` to implement the "latest
  object" query (providing an object ID and no version), using the
  checkpoint high watermark as the implicit checkpoint bound.
- `Epoch.denyList` to fetch the version of the deny list object as of
  the start of an epoch.

## Test plan

New unit tests in the reader for the new data loader:

```
sui$ cargo nextest run      \
  -p sui-indexer-alt-reader \
  -- test_checkpoint_bounded
```

New E2E tests for the GraphQL schema changes:

```
sui$ cargo nextest run         \
  -p sui-indexer-alt-e2e-tests \
  -- graphql/objects/at_checkpoint
```

The query underlying the data loader has also been performance tested on
the DB -- it was able to fetch 1000 object versions in 2 to 3s on a cold
database, and 4ms on a warm database.