indexer-alt: kv_packages (improved package indexing)
## Description

Adds support for improved package indexing, this includes:

- Detecting whether a packge is a system package or not, based on
  whether it was published/upgraded by a system transaction.
- Tracking all versions of system packages (not just the latest
  versions), to allow us to implement `Epoch.systemPackages`.
- Storing the serialized object for a package, rather than the
  serialized `MovePackage`, which allows us to serve package-related
  queries using just the `kv_packages` table (previously, `sum_packages`
  could only be used to serve package resolver queries, and the GraphQL
  schema related to packages was typically served from the larger
  object-related tables.

This change does not include use of this new table in any of our
readers, because it needs to be landed first and backfilled, before we
can start reading from it.

## Test plan

Indexing has been tested manually. Query performance has also been
confirmed as good, in particular, the fact that we are storing all
versions of the system package means we need to use a more complicated
query to fetch the latest version of a package (for package resolution
queries), similar to the queries we use on `obj_versions`. These were
tested by dumping the following from the production database, to
simulate `kv_packages` locally:

- `sum_packages`
- `obj_versions` for system packages

(Earlier versions of system packages were simulated by making a copy of
the contents of the latest version, but setting the correct version and
checkpoint).

Fetching the latest versions of ~80 packages (including all system
packages) on a semi-warm database (85% cache hit rate) completes in
1.9ms.