[faucet] Rework local faucet (#21865)
## Description 

This is a rather larger PR as it is best to commit these changes
together. There are two main changes here:
1) The faucet is now highly simplified by getting rid of the batching
code logic and making it a simple sequential faucet. You make a request,
you get a coin/coins (depending on how it's set up). This is a breaking
change as the response type has changed and it matches the one in the
devnet/testnet faucet.
2) All the places where the faucet is/was used are upgraded to use
`v2/gas`.

Importantly, this faucet will be only used for local network
deployments.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI:
- [ ] Rust SDK: