[sui][cli] Add ability to specify network in command
This enables the ability for `move` and `client` subcommands to accept a
`--env <network-name>` flag to use the specified `network-name` for the
single command (overriding the current active-env).

`network-name` must be a valid environment in the current config in
order to be used. E.g., a raw RPC URL for the `--env` name is invalid.