[sui][cli] Add ability to specify network in command (#21983)
## Description

This enables the ability for `move` and `client` subcommands to accept a
`--client.env <network-name>` flag to use the specified `network-name`
for the single command (overriding the current active-env).

`network-name` must be a valid environment in the current config in
order to be used. E.g., a raw RPC URL for the `--client.env` name is
invalid.


## Test plan 

Added test to make sure the `--client.env` flag properly overrides the
environment. The best way I could find to test this in a repeatable/sane
way was with `envs` and `active-env` calls.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [X] CLI: Add `--client.env` flag to `client` and `move` subcommands to
allow for selecting a specific environment for a single CLI command.
- [X] Rust SDK: Update arguments to `WalletContext::new` so it only
takes the path to the config. Additional configurations (timeout, max
concurrent connections, etc) can then be set on the context after it has
been created.

---------

Co-authored-by: Timothy Zakian <timothyzakian@mac.lan>