graphql: use sui-display to implement Display
## Description

Switch the implementation of `Display` in legacy GraphQL from a copy of
the fullnode implementation to the `sui-display` implementation.

This is a trial run for making the same change to fullnodes.

## Test plan

E2E tests:

```
sui$ cargo nextest run -p sui-graphql-e2e-tests
```

Changes affect output order, (entries are now ordered by key name), and
wording for errors.