Remove object size bounds on system transactions (#22173) (#22179)
Allows objects created/mutated by system transactions to not be bounded
by normal limits and brings it in-line with the rest of the way we treat
system transactions and limit interactions.

CI

---

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [X] Protocol: Allow larger objects to be created by system
transactions
- [ ] Nodes (Validators and Full nodes):
- [ ] gRPC:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [ ] Rust SDK: