Enable tidehunter with cfg rather then feature (#22221)
This means that even when cargo is run with `--all-features` it won't
enable tidehunter by default. This also allows for more concise check
for `#[cfg(tidehunter)]` in code.

With this PR to build with tidehunter one needs to run the following
```
USE_TIDEHUNTER=1 cargo --features typed-store/tidehunter <command>
```

## Description 

Describe the changes or additions included in this PR.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: