cli: Support setting a --gas-sponsor on tx commands
## Description

Make it possible to set the sponsor (gas owner) address on transactions
to something other than the sender's address. This makes it possible to
dry-run, inspect, serialize or compute the digest of such transactions.

There is also rudimentary support for executing a transaction with a
sponsor, but it requires the CLI to have access to the keys for the
sponsor, which is unlikely.

## Test plan

New CLI test, for a sponsored transfer.