Plumbing for accumulator events (#22296)
This PR contains most of the plumbing from #21701, which can be merged
more or less as-is.

Since this doesn't do anything on its own yet, the motivation may not be
clear in all cases, but #21701 can provide the context.

(Also, there are no tests yet because, again, this doesn't do anything
yet.)