[static-ptbs] Add ability to serialize values without needing layouts (#22375)
## Description 

We don't need a type layout when serializing a Move value in order to
produce valid serialized bytes.

This adds this function, renames the existing `simple_serialize` to
`typed_serialize`, and updates the usage of the the `typed_serialize`
function in the new static ptbs to use the serialization function that
does not require the type layout.

## Test plan 

Added an additional test for one interesting case, re-enabled
values-impl proptests and in this we ensure that the serialized bytes
for values are the same whether using `typed_serialize` or `serialize`.