[consensus] Return consensus position when calling submit_transaction (#22157)
## Description 

Removed logic related to getting effects from old `submit_transaction`.
New `submit_transaction` path will only submit transaction to consensus
and get back the consensus position. `wait_for_effects` will not submit
to consensus and will just check if effects are available to be
returned.

`ConsensusTxPosition` is based on the tx indices calculated while
transactions are being pulled when the block is being generated in
`try_new_block`. There is no guarantee that a returned
`ConsensusTxPosition` will be sequenced and it is up to the client to
retry and get a new consensus position if needed.

## Test plan 

Unit tests & simtest

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: