[move][move-compiler] Add modes to compiler (#22175)
## Description 

This adds modes to the compiler. Note this does not currently poison
non-test modes, but it is possible we should before landing it.

## Test plan 

New tests, plus existing ones.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------

Co-authored-by: Cameron Swords <cgswords@CameronacStudio.hsd1.wa.comcast.net>