[execution] add epoch stable sequence number to transaction effects for configs (#19439)
## Description 

This adds support for including the epoch-stable sequence number for
config objects (and the deny list object) in transaction effects. This
is useful in general, but particularly so for transaction replay.

This is done by adding a new marker value to the marker table that holds
the sequence number at which the object was mutated for any config
mutations or mutations to the deny list object. Importantly, there is
only one marker placed per-object per-epoch (this is ensured by always
placing the value at version 0, and query for the existence of the key
before placing the marker in the authority store). When we look for the
epoch stable sequence number for the object, we either return the value
in the marker table (if present) or the current version of the object.

This deprecates the old `UnchangedSharedKind::PerEpochConfig` and
replaces this with
`UnchangedSharedKind::PerEpochConfigWithSeqno(SequenceNumber)`.

This adds a new `ConfigStore` trait that enables accessing the per-epoch
stable sequence number for a given object.

This additionally fixes an error in `reconfigure_for_testing` where we
were not properly initializing randomness after a reconfiguration in
test scenarios.

## Test plan 

Added a unit test that checks that with different transaction orderings
(and mutations) the epoch-stable sequence numbers are fact stable.
Additionally updated the e2e `per_epoch_config_stress_test` to make sure
epoch-stable sequence numbers are stable as well.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [X] Protocol: Added epoch-stable sequence numbers for read-only
per-epoch configs accessed in the transaction.
- [ ] Nodes (Validators and Full nodes): 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: