[docs][ci] Action to support protobuf doc creation
Adds an action that triggers on a merged PR that includes changes to the .proto files. Auto-gens a PR that updates the JSON file that is processed by a component on docs site. Only opens PR if the JSON file has changed...which should always be true, but just in case.