indexer-alt: context for metrics address binding (#22512)
## Description

Add a context message on the error we get when failing to bind a metrics
address. This can fail if the address is already in use, and it's useful
to know what we were binding at the time.

## Test plan

Manually tested (run two instances of the indexer).

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: