[indexer-alt] Two table pruning implementation for obj_info and coin_balance_buckets (#22186)
## Description 

We can revert the obj_info and coin_balance_buckets from consistent to
simpler concurrent pipelines by changing the pruning approach. We
maintain a separate reference table, so that on commit, we write data as
usual to the main table, and in the same transaction write the keys of
those entries to the reference table.

On prune, we load the records from the reference table within the
pruning range, and use each entry to prune its immediate predecessor,
before deleting the reference record itself.

Note: object wraps/ deletions emit two records on the reference table.
One covers the previous modification, and the next one is always `cp+1`
for the current wrap or deletion record. This means that on pruning, we
don't remove the sentinel row at the cp it was deleted at, but the
immediate next checkpoint.

This also improves restartability - since the committer is responsible
for writing to both tables, we can restart from the
`checkpoint_hi_inclusive`, instead of `pruner_hi` today to regenerate
the in-memory cache.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: