[execution] Added missing support for PTB tracing (#22526)
## Description 

This PR adds support for the remaining PTB tracing in the execution
layer. I also modifies the debugger-side test to be more comprehensive
and include (almost) all command (upgrade command is not include but
it's the same as publish command).

The reason for modifying the test rather than creating a new one is that
these external tests take a fair amount of space due to all dependencies
from replayed transactions that they require

## Test plan 

A test has been modified to reflect additional PTB commands. All tests
must pass