indexer-alt: treat unrecognized configs as an error (#22566)
## Description

Previously, the indexer would warn on unrecognized configs, but would
keep executing. This posed a problem because GraphQL relies on the
indexer being faithful to this configuration (so it knows which
watermarks to read).

## Test plan

New unit tests:

```
$ cargo nextest run -p sui-indexer-alt
```

And manually tested:

```
$ cargo run -p sui-indexer-alt -- indexer --remote-store-url "https://checkpoints.mainnet.sui.io" --config config.toml
... [snip] ...
Error: Found unrecognized pruner field. This could be because of a typo, or because it was introduced in a newer version of the indexer:
prun-concurrency = 1
```

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: