indexer-alt: add uptime metric everywhere
## Description

Introduce an `uptime` metric to `sui-indexer-alt`,
`sui-indexer-alt-jsonrpc`, and `sui-indexer-alt-graphql`. This is
particularly helpful for determining the version and SHA of the service
that is currently running, per pod.

This works similarly to `mysten_metrics::uptime_metric` but was copied
into `sui-indexer-alt-metrics` instead because the duplicated logic is
minimal (adding a metric based on an elapsed time) compared to all the
other features that `mysten-metrics` offers, and it also does a couple
of things that we don't need/want (like including a chain identifier
label, and namespacing with a service name label).

## Test plan

Manually tested on each service, by running the service and inspecting
its `...:9184/metrics` output for an `uptime` metric (the metric itself
is namespaced by the registry it is attached to).