consistency(db): DbMap takes Arc<Db>
## Description

Have the `DbMap` take an `Arc<Db>` instead of a reference to a database,
and have it fetch the column family on every operation, instead of
caching it. Both these steps simplify the lifetime story for `DbMap` and
`Db`.

Also changes the argument order to `Db`'s constructor. This is an
unrelated change, but touches the same code, and was done to put the
`path` of the Db first, as it in some ways identifies the database.

## Test plan

Existing tests