consistency: Store implementation
## Description

An implementation of the indexing framework's `Store` trait, on top of
`Db` and `DbMap`. The implementation only supports the transactional
setting (without pruning), and writes work by sending data and a
watermark down a channel to a task that coordinates with other tasks to
land writes to the database and take snapshots.

## Test plan

New unit tests:

```
$ cargo nextest run -p sui-indexer-alt-consistent-store
```