Make `replace` in `TxContext` Move "variadic" (#22701) (#22724)
## Description 

`TxContext::replace` was changed in the number of args from 8 to 9. We
thought that was ok given implicit deps and the test only function. It
turns out that because of the release schedule and the multiple versions
of the CLI that can be used having both version work would be ideal,
maybe necessary.
This PR makes the native work both when 8 or 9 args are provided.
Basically the native works for both the old and the new version of the
framework

## Test plan 

Tested manually on the cli run that fails

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

## Description 

Describe the changes or additions included in this PR.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: