[trace-view] Fixed a problem with updating Global locations (#22824)
## Description 

This PR fixes a problem related to handling `Global` trace values in the
trace debugger. A bit of context first, the way the debugger handles
references is that it creates a chain of local variables (containing
references) that ultimately points to the location where the actual
values is stored. When reading/writing the value off a given local, this
chain is traversed and the original value is read/written (and it's the
only place where the things are updated - locals containing references
are not).

There are two types of reference chains:
- one that ends in a "regular" value (e.g., a struct)
- one that ends in a `Global` value

Global values are references that appear in Move code execution "out of
thin air" (e.g., a mutable reference is passed as an argument to the
top-level function). If not handled correctly, this may create a problem
when assigning values to globals.

In the debugger's runtime, global values are stored in in a map
(initially populated by `DataLoad` effect which initializes it from the
global's value snapshot). However, subsequently, there may be a
situation when a reference is assigned to a global in the trace. For
example:
```
{"Effect":{"Write":{"location":{"Indexed":[{"Indexed":[{"Global":16},0]},0]},"root_value_after_write":{"MutRef":{"location":{"Global":16},"snapshot": ...
```
What's important here in this trace snippet is that we are assigning a
reference to `Global16` (which was just modified) to a variable
containing a reference path to the same `Global16`. Prior to this PR, we
would process this trace effect directly and assign reference to
`Global16` to be the value of `Global16`, creating a global that
effectively points to itself (more precisely, to its location in the
map). When subsequently trying to read a variable pointing to this
global, we would end up in infinite recursion trying to access the end
of the reference chain.

This PR fixes this problem by always assigning a value to globals from
(already updated) snapshots), much like it happens when they are
initialized.

When working on this PR, a problem with the trace generation was
discovered, which was subsequently fixed in
https://github.com/MystenLabs/sui/pull/22777. Variable values present in
the output of the new test added to this PR reflect the updated
(correct) global value tracking in the fixed trace generation problem.
While not strictly necessary it is thus probably best to land this PR
after the one containing the fix lands

## Test plan 

Additional test was added that would fail to correctly display variable
value after assignment of a global to a an indexed global. All tests
must pass