consistency: expose stride through AvailableRange
## Description

The stride is the gap between checkpoints at which the service takes
snapshots. If GraphQL is combining data from a consistent store with
other stores, it needs to know how often the store takes snapshots in
case another store is lagging behind the consistent store, and it needs
to re-align all data on some earlier checkpoint.

## Test plan

Manually tested.