New indexer arg '--task' to enable pipelines to run with watermark entries keyed separately from main pipelines. Take out skip_watermark on the top level. Separate task that periodically checks value of reader_lo from main pipeline.
Make delimiter part of the Store trait. The choice of delimiter is only relevant to the implementor of Store. At the indexer level, an operator can just choose to pass in an optional task string. The exact watermark_key to be created is thus up to the Store implementor, and otherwise invisible to the indexer operator. Should also make it easier to validate pipeline names going forward. Address nits and other smaller comments