Use package version ID when reading debug info (#23051)
## Description 

This PR implements improvement to source-level debugging for replayed
transactions. The limitation of the previous approach was that there was
no way to source-debug upgraded packages. The reason for it was that we
were relying on the build process to populate debug info with the
correct package ID and there was no way to make this package ID
represent upgraded `version_id` (we could only a package version to
represent `original_id` of the non-upgraded package).

In this PR we re-organize structure of the directory created by the
replay tool to contain trace-debugging meta-data to make `source`
directory a sub-directory of a package directory (named after package
ID). We can then override package ID in the debug info based on the
package ID taken from the directory name rather than relying on the
build system to do it for us.

## Test plan 

Manually tested that source-level debugging works in the new setup