Settle balance after executing settlement transaction (#23013)
## Description 

Call the settle balance API on the balance withdraw scheduler after
committing a settlement transaction.
This PR refactors checkpoint builder a bit so that we keep the
settlement information when building out the settlement transaction.
That information is then later used to call settle balance.
This also allows us to simplify settle balance interface a bit because
we now can guarantee that this is called in order, once per accumulator
version.

## Test plan 

Verified that some simtests start to pass.
---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: