build file system and in memory caches (#23229)
## Description 

This is the code for the file system store, the in memory store and a
read through store.
Combinations of those store should allow for different kind of efficient
runs.
I have tested manually that the caches and their combination work as
expected.
This should also allow for replacing of packages and run of the same
transaction with the update packages. I have not tested that yet.
Next steps are:

testing including the creation of a local cache (file system) that can
run some proper testing for the tool
review of the command line options vs config file. I am thinking we
should use both to simplify the command line
interesting transactions reviewed, run, saved and used for testing (not
the replay tool but the core system testing)
As a warning: the code hangs when running from a file with more that a
handful of transactions. That is not introduced by this change but it's
there in main already. It hangs inside the network request. I am not
sure yet as to why. Played with timeouts but that does not seem to be
the problem. I will investigate more

## Test plan 

Manual test for now

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: