refactor(graphql-alt): Object/MoveObject share `OnceCell`
## Description

`OnceCell` supports clone, but the clones don't share the same contents,
so if one clone gets initialised, the contents are not propagated to the
other clone. This change wraps the `OnceCell`-s used for `Object` and
`MoveObject` in an `Arc` so that they do become shared (and removes a
wrapping by `Arc` nested more deeply in the type).

## Test plan

Existing tests:

```
$ cargo nextest run -p sui-indexer-alt-graphql
$ cargo nextest run -p sui-indexer-alt-e2e-tests -- graphql
```