[docs] update wallet standard docs to be more technically accurate (#23244)
## Description 

This PR updates our Wallet Standard documentation to reflect our new
recommendations/patterns. For a summary of changes:
- Updated the documentation around how account authorization works
(based on
https://github.com/wallet-standard/wallet-standard/pull/128#discussion_r2056842722).
Generally, I think using the terminology `connect` is confusing and
leads to misunderstandings of how the standard works at a technical
level
- Removed the documentation around `reportTransactionEffects` since we
are going to deprecate it and don't recommend using it
  - Fixed various code typos and syntax errors

JFYI: Some of this is AI generated with manual edits and tweaking so no
offense taken for any writing feedback

## Test plan 
-
https://sui-docs-7xvpqgady-sui-foundation.vercel.app/standards/wallet-standard
👁️

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------

Co-authored-by: ronny-mysten <118224482+ronny-mysten@users.noreply.github.com>