refactor(graphql-alt): MovePackage lazy loading
## Description

Support lazy loading the native form of `MovePackage`, similar to the
set-up for `MoveObject`. This is not leveraged in this PR, but it will
allow us to represent a `MoveModule` or a `MoveFunction` by its
fully-qualified name, without incurring a DB fetch to make sure the
package/object exists.

## Test plan

Existing tests:

```
$ cargo nextest run -p sui-indexer-alt-graphql
$ cargo nextest run -p sui-indexer-alt-e2e-tests -- graphql
```