[coin registry] Add internal rule for `new_currency` (#23377)
## Description 

Adds an internal rule for `new_currency`. system call, as we want `T` to
be internal.

## Test plan 

Snapshot tests for both the compiler & the execution

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: