[bella-ciao][sui-adapter] Rough-in of new adapter on new VM
This roughs in the new adapter on top of the new VM. There are still
some TODOs, and changes that need to be made -- in particular the value
representation and loading from the adapter into the VM currently uses
the `StackFrame` API (which is wrong) and this should be moved over to
use the `BaseHeap`. Additionally we don't support function calls into
the VM yet because the VM doesn't have the deserialized value function
call API (yet).

Additionally there are some cleanups that can be performed on the
adapter side now such as I believe `RootedLinkage` can be removed
entirely in favor of `ResolvedLinkage` which should make some things a
bit cleaner.

There were some additions to the VM's API, in particular there is now an
API that allows you to get type information out about a VM Type
(original ID, defining ID, abilities, etc). This is needed for the
adapter code.