fix(errors): Don't use debug format for errors
## Description

`SuiErrors` are user-facing. As much as possible, we should avoid using
`Debug` formats for the constituents of these errors, because they may
make it to developers.

This also fixes a CI breakage in `main` due to an error being wrapped in
another error that displayed it using its Debug format.

## Test plan
CI