[feat] allowing submitting latency check requests
[review] address review comments. Extend wait_fo_effects response to include details when requested.

[review] address review comments. Fix failing tests.

[feat] add the MovingWindow component.

[refactor] create the moving_window for average stats calculation. Remove the dacay_moving_average. Refactor the code to support the new component.

[refactor] do not use the global stats. Calculate the max when the scores are getting updated.

[refator] measure only the finalisation (end to end) latency. Tune weights to take into account only reliability, health check, and finalisation latency into account.

restore decay_moving_average

[fix] failing test. Address review comment.

[fix] linter.

[MFP] Introduce the ValidatorClientMonitorPool (#23544)

Enhancing the `ValidatorClientMonitor` to calculate and hold the scores
per transaction type separately. In practice all the metrics are reused,
the only thing that differs is the `FastPath` and `Consensus` end to end
latency operations that are taken into account when the corresponding
(by tx type) scores are calculated.

CI

---

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] gRPC:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [ ] Rust SDK:

[feat] Run periodic latency checks using the transaction driver.

[fix] remove unnecessary file.

[chore] extend metrics.

[refactor] ping all the validators. Decrease the frequency of latency checks.

[fix] RequestRetrier could panic when an allowed_list validator not found in the authority aggregator clients.

[refactor] small refactor

[chore] reduce the moving window size.