[cherry-pick][cli] Move the protocol config loading into the flag branch  (#23472)
## Description 

This PR fixes the issue for upgrade compatibility checks which appears
when trying to do an upgrade with an older CLI than the current protocol
version on the network, but without passing the `--verify-compatibility`
flag. The code is moved in the if branch as `--verify-compatibility` is
false by default.

## Test plan 

Existing tests.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [x] CLI: Fixed a bug where an upgrade command would terminate early if
the CLI binary is not at the same protocol version or newer than the
network.
- [ ] Rust SDK: