[1/n][bella-ciao][move-vm] Update native extensions to be cloneable (#23450)
This wraps the `NativeContextExtensions` in an `Arc<RwLock<` to allow
cloneability and interior mutability. We could probably use a `RefCell`
or `Mutex` instead of an `RwLock`, but figured this was the most in
keeping with what we have elsewhere, and the most "safe" option. Happy
to take feedback on it being something else though.