graphql-alt: Object.dynamicField et al
## Description

Add dynamic fields to `Object`. Previously these functions existed on
`Address` and on `MoveObject` but not on `Object`, but I spotted based
on the example queries that we did used to have it there.

The reason it's not part of the `IAddress` interface is that this would
require us to implement it on `MovePackage` where it would be entirely
redundant.

## Test plan

Updated E2E tests:

```
$ cargo nextest run -p sui-indexer-alt-e2e-tests -- graphql/objects/dynamic_field
```