CoinRegistry/GRPC Integration (#22874)
## Description

Integrating GetCoinInfo gRPC API with CoinRegistry system object.

For reference, the CoinRegistry Move implementation is
[here](https://github.com/MystenLabs/sui/pull/21400/commits/589be6f935394cf2e7cb0cc10cbb9c9a5d5f3942).

Note that for CI to pass -

-
[MAX_PROTOCOL_VERSION](https://github.com/MystenLabs/sui/blob/9e905744303665ae6ed62ff71b75d000f3e83fba/crates/sui-protocol-config/src/lib.rs?plain=1#L22)
must be bumped to 89 (not included in this PR).
- Proto definition updates need to be merged (PR
[here](https://github.com/MystenLabs/sui-apis/pull/3)).

GetCoinInfo first checks the CoinRegistry dynamic fields and uses the
CoinData field for the request coin type if it exists, otherwise it
falls back to the CoinMetadata in the coin index.

Things to Note
* I made a change to call the coin's supply "fixed" if the treasury cap
is owned by 0x0 as discussed in slack. We can add more logic to this
function to handle more legacy supply freezing cases as desired.
* I extracted a utility function out of the balance API's e2e test to
publish move packages to the test cluster, which is why you'll see a
diff in that test.
* I updated the `wait_for_transaction` utility function to wait until
the transaction has been included in a checkpoint, not just validated by
validators to fix some race conditions in the tests.

## Test plan

New unit tests that create and manipulate coins with coin reigstry APIs
and assert the gRPC response contains the expected metdata.

---

## Release notes

- [x] gRPC: Integrating GetCoinInfo with new CoinRegistry system object.