[TransactionOrchestrator] remove effects subscription (#23590)
## Description 

This logic is only used in tests, and it is not the best way to return
execution effects (already returned via `execute_transaction_block()`).
`QuorumDriver` has this mechanism too, but it seems only useful for
cleaning pending txn logs on recovery, which can be implemented in
another way too.

## Test plan 

CI