[5/n][bella-ciao][values] Remove old global fingerprint for values and use the object runtime's fingerprints (#23513)
## Description 

@tnowacki added global value fingerprints to the object runtime a while
ago in anticipation of this. This completes the job by removing the
hacky global fingerprints that we used as an interim solution in the VM.

This removes mutation tracking for global values from the VM values (and
instead this is the underlying responsibility of the runtime). This
allows us to simplify the internal representation of global values so
they are no longer a four-state value but instead a two-state (empty or
filled).

## Test plan 

Move tests are green. 


---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: