[replay] Added config file for default option values (#23578)
## Description 

This PR adds the ability to add config file for the replay tool which
would hold default values of the options otherwise specified on the
command line. The command-line options take precedence over the ones
specified in the config file.

The implementation, somewhat unfortunately, requires re-parsing command
line arguments when replay tool is used from Sui CLI. Happy to hear
suggestions on how to avoid it

## Test plan 

Tested manually