[MFP] Introduce the ValidatorClientMonitorPool (#23544)
## Description 

Enhancing the `ValidatorClientMonitor` to calculate and hold the scores
per transaction type separately. In practice all the metrics are reused,
the only thing that differs is the `FastPath` and `Consensus` end to end
latency operations that are taken into account when the corresponding
(by tx type) scores are calculated.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: