bugfix(graphql-alt): allow `null` variables for optional parameters
## Description

Previously the query limit checker would complain if a variable was
referenced in a query but no binding was provided for it, regardless of
whether the variable's type is optional or not.

This can cause the service to reject valid queries, like:

```
query ($first: Int) {
  transactions(first: $first) { nodes { digest } }
}
```

...if no variables are supplied.

This change relaxes this requirement by returning `null` if the variable
is missing, rather than an error.

## Test plan

Updated tests:

```
$ cargo nextest run -p sui-indexer-alt-graphql
```