bugfix(graphql-alt): allow `null` variables for optional parameters (#23636)
## Description

Previously the query limit checker would complain if a variable was
referenced in a query but no binding was provided for it, regardless of
whether the variable's type is optional or not.

This can cause the service to reject valid queries, like:

```
query ($first: Int) {
  transactions(first: $first) { nodes { digest } }
}
```

...if no variables are supplied.

This change relaxes this requirement by returning `null` if the variable
is missing, rather than an error.

## Test plan

Updated tests:

```
$ cargo nextest run -p sui-indexer-alt-graphql
```

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [x] GraphQL: Fix a bug where queries would fail if they used a
variable to populate a nullable parameter, and then did not supply the
variable (which is a valid thing to do).
- [ ] CLI: 
- [ ] Rust SDK: