[9/n][bella-ciao] Remove `effects` from `move-core-types`
We were no longer using this at all in the new VM, and it did not make
sense to keep it around and would be a hidderence to some upcoming
changes where we want to have `CompiledModule`s at the boundary instead
of bytes.

This is a also a nice cleanup both  on the VM side, and also in the
object runtime.