[move-ide] Incremental symbolication fix (#23666)
## Description 

This fixes a problem with incremental symbolication. What we attempted
before was to collect information about which files are modified (i.e.,
which ones have to be fully compiled) from the messages sent by the
client to the LSP server. In some cases, however, this seemed to have
led to problems.

In particular, a file might have been deemed unmodified by the analysis
but be actually modified before mapped files structure (including file
hashes) used throughout analysis was constructed. This would result in
the discrepancy in the hash for this file seen by the compiler (as it
would used cached data for this file) and mapped files structure (as it
would have a hash after modification). This could cause `move-analyzer`
to crash.

This PR implements a much more robust way of determining which files are
modified that manages to introducing this race condition

## Test plan 

This was difficult to reproduce, but the original reporter of this bug
confirms that after trying `move-analyzer` with this fix, the crashes he
used to observe no longer happen