[sui-pkg-alt] fix move-cli and move-analyzer tests (#23637)
Update the tests in `move-analyzer` and `move-cli` to work on the
`sui-pkg-alt` branch. Some tests were disabled (for features we're
removing), some were deferred (because we need, e.g. modes), and some
were modified in hopefully harmless ways (e.g. updating snapshots). Here
are the notes I was keeping (they may be a little out of date)