display: new syntax for dynamic field access.
## Description

Differentiate the syntax for vector access from the syntax for dynamic
field and dynamic object field access so that we can always determine
from the syntax which accesses are going to load objects. The new syntax
is as follows:

```
e[f]    # vector or VecMap access
e->[f]  # dynamic field access
e=>[f]  # dynamic object field access
```

In future, we may also supporting "dereferencing" an object by its ID,
and fetching a field from that value, with the following syntax:

```
e->f
```

And it is TBD what the syntax would be for loading a value, wholesale.

## Test plan

```
$ cargo nextest run -p sui-display
```