[sui-adapter] Simplify event call check in execution (#23804)
## Description 

Didn't catch this in #23182 but we don't actually need to add the check
against the `PRIVATE_EVENT_FUNCTIONS` and could have kept the existing
checks that were there which are a bit more robust (as they ban any
calls into the event module not just specific functions even though
those are the only ones you can call today), so this moves it back to
the pre-existing check.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: