[17/n][bella-ciao] Variable renaming of `runtime_id` and `storage_id`
Go through the VM and adapter to make sure that we consistently use
`version_id` and `original_id` instead of `storage_id` and `runtime_id`.

Also check that the type aliases that we use in the VM for
`OriginalId`/`VersionId` and `DefiningTypeId` all match up with what
they should be using.

No semantic changes, should only be renames of variables/types