[MFP] allow list for tx submission (#23853)
## Description 

Set an allow list to submit transactions via transaction driver to
specific validators only. The provided list in the configuration is
using the validator display names, as this is quite straightforward for
someone to setup instead of dealing with serialised keys/addresses etc.
The solution does not update the list when the committee updates, but
this could be easily done if required. For now opted in for a simpler
solution.

## Test plan 

CI/PT

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: