display: string hardening
## Description

Be more paranoid about string values: handle them as raw bytes in the
interpreter, and perform a checked conversion to `&str` if we need to
output it as a string.

This is to handle a case where the serialized root object is controlled
by the user, and can be used as a way to introduce invalid UTF-8 through
BCS deserialization.

Note that invalid UTF-8 values that show up in string literals were
already checked for by the parser.

## Test plan

New unit tests:

```
$ cargo nextest run -p sui-display
```