[cli] Add a mechanism to cache the chain id to the `client.yaml` file. (#24072)
This PR adds a caching of chain-id to the Sui CLI. This will be used in
the new pkg-alt system.

In particular, the caching will happen when one of the `client` commands
will do an RPC request anyway / or execute a tx. It will not happen when
any client command does not work with the RPC.

For `publish` and `upgrade`, it will try first to load it from cache; if
it does not exist, then it will cache it. For all other commands it will
always cache it.

Manual test with check on the `client.yaml` file. Tested also the
updated `client.yaml` file with an old CLI to ensure backwards
compatibility.

---

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] gRPC:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [ ] Rust SDK: