Increase consensus client timeout (#24104)
## Description 

This timeout is hit when submission to consensus becomes impossible when
the validator transitioned into a full node, but the submission is not
cancelled. The fundamental fix is to cancel the submission properly on
epoch change. Increasing timeout is a short term mitigation.

This timeout can be made to be infinite, but that can hide issues. So
using a longer but finite timeout.

## Test plan 

CI