Add accumulator version to assigned versions (#24146)
## Description 

Previously accumulator version is in a separate field, where only if a
transaction has address balance withdraws, we attach the accumulator
version. However to support object balance withdraws, we will need it
even when we don't see address balance withdraw reservations.
This PR moves this field out of the separate field and make it part of
assigned versions, since it technically comes out from version
assignment. It will be set as long as balance accumulator is enabled.
The primary change in this PR is that now we have to also calculate the
accumulator version in the checkpoint executor for each transaction,
since we don't conveniently have that info like consensus.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: