[package-alt] turn on info output for move_package_alt crate (#24110)
Adds a new tracing subscriber that logs unadorned info output
for a given set of crates, and enables it for the move_package_alt
crate. This will show the output from the git and external resolver
commands by default.

I think this is a stopgap; ideally we would do a better integration of
the output in different places in the codebase, but it seems like the
fastest way to get the output from the move_package_alt crate.