Return missing instead of internal error for checkpoints that haven't been fully stored yet (#24244)
## Description

In the indexer logs we see `Retrying due to error: Failed to fetch
checkpoint 211394366: status: 'Internal error'` at the tip-of-chain when
a checkpoint hasn't been fully stored yet. This should be a "Not Found"
rather than an "Internal Error." This is especially important because
the indexing framework retries with exponential back off on internal
errors but linear back off on "Not Found" so returning an Internal Error
is actually increasing lag.

## Test plan

None really 🙈. It's kind of tough to write a unit test because this
requires calling the API at the exact instant where the checkpoint
exists but the effects haven't been stored yet.

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [x] gRPC: Return "Not Found" for new checkpoints that haven't been
fully stored yet instead of "Internal Error."