[package-alt] split system dependencies and implicit dependencies (#24156)
Split up implicit system deps into implicit deps and system
deps.

There is a new dependency type called `{ system = "name" }`. The
`system-dependencies = [...]` option is replaced by an
`implicit-dependencies = false` option.

The flavor defines a set of implicit dependencies and a set of system
dependencies. The implicit dependencies are manifest deps that are added
as long as implicit-deps is not false. The system dependency table is
used to resolve the `{ system = "name" }` deps.