Implement obj_versions pruning without overhead.
use sui-sql-macro query, through sui-pg-db. seems like we have to manually import sui-pg-db because the path is hardcoded in the macro itself. lost my work - forgot to git stash, but managed to restore it. Initial version is a single query using multiple ctes, with inline comments. more testing to follow