[consensus] fix deadlock during recovery (#24292)
https://github.com/MystenLabs/sui/pull/24024 introduced a potential
deadlock when recovering consensus. Possible sequence of events:

1. Thread A: acquires `dag_state` read lock for the duration of the
function call in
`recover_blocks_after_round(observer.dag_state.read().gc_round())`
2. Thread B: `CommitFinalizer::run()` tries to acquire the `dag_state`
write lock, and is blocked on the reader lock from (1)
3. Thread A: `recover_and_vote_on_blocks()` tries to acquire read lock
on `dag_state` again. This usually succeeds but in this case because of
(2), this is blocked, causing a deadlock.

Also:
- Run the last step of `TransactionCertifier` recovery in a blocking
thread.
- Improve logging.

CI
The affected validator is fixed.