Certora audit fix: L-07 (#24282)
## Description 

This is a fix for issue L-07 from the Certora audit of the native
bridge.

The audit can be seen
[here](https://www.notion.so/mystenlabs/Certora-Audit-2966d9dcb4e980c490e2ee96b63dbe82?source=copy_link).

> The delay function is supposed to double the delay time every attempt.
However, instead of using the power function it uses an xor with the
number of the attempt. This would lead to a weird series that’s similar
to linear (3,0,1,2, 7,4,5,6 etc.)

In rust `^` is the XOR operator, not exponent.

## Test plan 

Nothing other than CI tests.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] Indexing Framework: