Fix RefCell borrow (#24390)
## Description 

We were trying to insert into the object cache while holding a reference
(`borrow`) to the cache, resulting in an "already borrowed" error when
trying to get the mutable reference to insert

## Test plan 

local testing of heavy transactions

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] Indexing Framework: