refactor(indexer): CancellationToken -> Service
## Description

Avoid using `CancellationToken` to detect when a service is being
shutdown while yielded, across the indexing and RPC codebase. When using
the `Service` abstraction, explicitly waiting on a cancellation token is
not necessary, as the yield will be within the scope of a task
maintained by the service which can be aborted.

As part of this change, Chain Identifier initialization was adapted.
Previously, the chain identifier task ran before RPC started, which
complicated lifecycles, because we needed to run that task to completion
to get a value before the service could start, which also meant that we
needed to handle termination signals for this initialization task as
well as for the main service.

Now, the chain identifier task is a secondary service that writes the
chain identifier to a `SetOnce` when it's done. This can be started
alongside the RPC, and the requests that need this information will wait
for it to be set before they can continue (but the RPC as a whole can
start and serve other requests before then, and termination handling can
remain in one place).

In a recent change the commit watermark was made to perform an extra
iteration before shutting down, however this can cause stalls during
shutdown if the task fails to connect to its store.

This situation occurs in calls to `sui start` where the CLI is
responsible for running the database process as a child as well as the
RPC services. The database is wound down by the `Ctrl-C` signal, while
the RPC services are gracefully shutting down.

The change was originally made to cater to the checkpoint indexer (or
any use of the object store) where watermark updates are infrequent, so
it was useful to flush the watermarks before shutting down in response
to a cancellation.

However, at the time of that change, the cancellation token was
responsible for signaling "normal shutdown" (e.g. after having processed
all checkpoints), as well as interrupts (unexpected termination).

The hope is that by using separate methods to communicate these two
forms of exit, we can avoid adding this graceful shutdown support.

## Test plan

CI -- this is a behaviour preserving refactor.