wip(refactor(reader)): CancellationToken -> Service
## Description

Avoid using `CancellationToken` to detect when the service is being
shutdown while yielded. When using the `Service` abstraction, explicitly
waiting on a cancellation token is not necessary, as the yield will be
within the scope of a task maintained by the service which can be
aborted.

Also make use of `Service` for the system package task.

## Test plan

This change breaks the build overall, but the crate builds and tests:

```
sui-indexer-alt-reader$ cargo check
sui-indexer-alt-reader$ cargo nextest run
```