wip(refactor(graphql)): CancellationToken -> Service
## Description

Switch GraphQL-alt to using the new `Service` abstraction, and away from
using `CancellationToken`.

As part of this change, Chain Identifier initialization was adapted.
Previously, the chain identifier task ran before RPC started, which
complicated lifecycles, because we needed to run that task to completion
to get a value before the service could start, which also meant that we
needed to handle termination signals for this initialization task as
well as for the main service.

Now, the chain identifier task is a secondary service that writes the
chain identifier to a `SetOnce` when it's done. This can be started
alongside the RPC, and the requests that need this information will wait
for it to be set before they can continue (but the RPC as a whole can
start and serve other requests before then, and termination handling can
remain in one place).

## Test plan

Build, test and lint `sui-indexer-alt-graphql`. Also run the service
locally and ensure that it exits cleanly when asked to shutdown, or when
forcefully aborted.