[move][move-compiler] Provide name suggestions when resolution fails (#24016)
## Description 

Introduces suggestions for module and alias name resolution in the Move
compiler when naming fails.

Suggestion is via a filtered list of candidates, with a hand-tuned
Levenshtein implementation to try to be somewhat performant when looking
for a candidate.

There are a lot of heuristics in place around when to suggest a name and
what kind, which I am sure can use tuning and adjustment based on usage
experience and feedback.

### Smaller changes:

* Improved error location reporting in CLI output for duplicate module
definitions, giving more precise spans in `args.exp`.
* Refactored `aliases.rs` to add clearer sectioning for type
definitions, traits, and implementations, and imported additional types
for improved clarity and maintainability.
* Added a convenience method `module_name` to `ModuleIdent_` for easier
access to module names.

## Test plan 

Tests were changed, and new output was hand-reviewed. Some specific
suggestion tests were added.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: