wip(refactor(consistent-store)): CancellationToken -> Service
## Description

Migrate the consistent store crate to using the `Service` abstraction,
away from using cancellation tokens.

This is mostly a straightforward conversion, with the exception of the
finalization process after restoration, which needs to run after the
main restoration service has exited successfully. Initially, I tried
adding finalization as a concept to the `Service` abstraction, but it
was not a good fit, because an individual service's finalizers would end
up dependent on the success of unrelated service's primary tasks after
merging.

Instead, it was easier to have the restoration task return a finalizer
that can be run to get its own service, which will run if the main
service finishes successfully.

## Test plan

Crate's own build, tests, lints.